# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Mobile",
        "description": "API Mobile",
        "contact": {
            "email": "marcos@agr-sistemas.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/cidade": {
            "get": {
                "tags": [
                    "Cidade"
                ],
                "summary": "Lista de Cidades",
                "description": "Retorna uma lista de cidades",
                "operationId": "getCidade",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cidade/{id}": {
            "get": {
                "tags": [
                    "Cidade"
                ],
                "summary": "Pega as informações da Cidade",
                "description": "Retorna os dados da cidade",
                "operationId": "getCidadeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Código",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cidade"
                ],
                "summary": "Exclui uma Cidade",
                "description": "Excluir a cidade",
                "operationId": "delCidade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Código",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/uf": {
            "get": {
                "tags": [
                    "UF"
                ],
                "summary": "Get list of ufs",
                "description": "Returns list of ufs",
                "operationId": "getUf",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UF"
                ],
                "summary": "Post a pet in the store with form data",
                "operationId": "postUf",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Uf"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "405": {
                        "description": "Invalid input"
                    },
                    "422": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/uf/{id}": {
            "get": {
                "tags": [
                    "UF"
                ],
                "summary": "Get uf information",
                "description": "Returns uf data",
                "operationId": "getUfById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Código id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "UF"
                ],
                "summary": "Updates a pet in the store with form data",
                "operationId": "putUf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of uf that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Uf"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "405": {
                        "description": "Invalid input"
                    },
                    "422": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "UF"
                ],
                "summary": "Get uf information",
                "description": "Returns uf data",
                "operationId": "delUf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Código id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Segmento": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "descricao": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Uf": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uf": {
                        "type": "string"
                    },
                    "descricao": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Acess token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}